+++ Sources/BaseFoundationiOS/Common/GlobalTask.swift
@@ -1,0 +1,17 @@
+//
+//  File.swift
+//  
+//
+//  Created by pablo borquez on 20/01/24.
+//
+
+import Foundation
+
+public class GlobalTask {
+    public static func performDelay(seconds: Double = 0.1, handler: @escaping SimpleHandler) {
+        Task {
+            try await Task.sleep(for: .seconds(seconds))
+            handler()
+        }
+    }
+}